+++ src/fr/maxlego08/menu/inventory/inventories/InventoryDefault.java
@@ -54,8 +54,9 @@ public class InventoryDefault extends VInventory {
 
         Collection<Pattern> patterns = this.inventory.getPatterns();
 
-        this.buttons = this.inventory.sortButtons(page, args);
+        this.buttons = new ArrayList<>();
         this.buttons.addAll(patterns.stream().flatMap(pattern -> pattern.getButtons().stream()).collect(Collectors.toList()));
+        this.buttons.addAll(this.inventory.sortButtons(page, args));
         this.buttons.forEach(button -> button.onInventoryOpen(player, this));
 
         this.updatedButtons = this.buttons.stream().filter(Button::updateOnClick).collect(Collectors.toList());